This is updataed version, which is debugged and runnable. But I`d like to improve it with more functional specifications as the current version just meets conceptual purpose of the issue.
